Create Spinnable for sharing common functionality
There are many common functionality for spinner and squad
